Add basic compatibility with browserify and webpack
On the compile step, both window and module are defined. Gun prevents module exports by detecting the environment and tacking Gun onto the window object. This addition will check to see if it's a dual environment (like browserify or electron) and export to both window and module, preventing errors in your build step.