Make data.json human-readable
Since the file.js module is only for development or getting started, we
shouldn't have to minify the data before persisting it. In fact, it may
hurt things, since it obfuscates our data structure and makes it harder
to understand what's going on under the hood.
In this commit, I swapped `Gun.text.ify` with JSON's `stringify` and
gave it a 2-space formatting option to make the data more readable.