Gun in a web worker = more speed, storage
I simply changed all the references of `window` to `self`. This is because self denotes a global variable inside web workers. I'm surprised that I was able to read from indexedDB and also write to indexedDB from inside the web worker. 

Web workers are helpful for improving page speed, which is why I wanted Gun to have a web worker implementation so much. I'm going to refactor my site to postMessage() to web workers instead of interacting with the DOM directly. woohoo